ml-service/requirements.txt
# ML Service Dependencies
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
lime==0.2.0.1
joblib==1.3.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
